feat: 스터디 정렬 로직 추가 - IN-PROGRESS -> UPCOMING -> ENDED 순 - 각 범위 안에서는 점수가 높은 순서대로     - 점수는 커리큘럼 진행도(%) + 학습자료 개수로 계산

+++ src/main/java/doore/study/domain/repository/ParticipantCurriculumItemRepository.java
@@ -19,7 +19,7 @@ public interface ParticipantCurriculumItemRepository extends JpaRepository<Parti
 
     void deleteAllByCurriculumItemId(Long curriculumItemId);
 
-    long countByCurriculumItemIdIn(List<Long> curriculumIds);
+    int countByCurriculumItemIdIn(List<Long> curriculumIds);
 
-    long countByCurriculumItemIdInAndIsCheckedTrue(List<Long> curriculumIds);
+    int countByCurriculumItemIdInAndIsCheckedTrue(List<Long> curriculumIds);
 }
